Hej, the json helpers file is now copypasta the cbor one, just different package name.
This is a feature, not a bug :)  All the interesting stuff is hidden -- correctly! -- behind the Decoder/Encoder types and their factory methods.

Punting on adding a `MarshalIndent` helper at the moment.  Glue your code together with the stdlib one if you need it.  (I looked at just vendoring that feature, but it's tied up with their scanner machinery, and I don't wanna deal with untangling that.  We can add this feature ourselves by adding proper indent support to our encoder state machine, where it belongs.)

Signed-off-by: Eric Myhre <hash@exultant.us>