Benchmarks over legacy systems gone.
I can't maintain that anymore, and the period of utility has passed.

The entire objLegacy package can probably be discarded now, but that's not the point of this branch, so we'll leave it be for now.  The tests, though, are actively too much to maintain as of this next commit.

Signed-off-by: Eric Myhre <hash@exultant.us>