Implement new top-level porcelain helpers!
You can now `refmt.Marshal(json.EncodeOptions{}, someObj)` and it will do the right thing -- and you can switch encoding format by value!  This should be a nice ease-of-use thing if you have an API where, say, some URL parameter or suchlike lets you change the serialization format on the fly.

The actual *use* of those Options structs is a bit forward-looking... as you can see, in this commit we're not actually introducing them into the various Encoder and Decoders at all, because as of this moment and branch, we have no internal features for them to drive.  But it's enough to get these top level helpers to have the right shape of params, so, mission accomplished.

Signed-off-by: Eric Myhre <hash@exultant.us>