Quick but critical fix to unmarshallers that was introduced by the porcelain refactor.
Kind of necessary to pump the decoder before figuring you can hand a token to the unmarshaller, eh?

This bug was made possible by the fact that the TokenSink and TokenSource interfaces are actually the same stepfunc definitions.  Perhaps we should change that :I

Added test coverage at top level to cover this kind of last mile thing, which should also help prevent any high-comedy regressions like this in the future.

Signed-off-by: Eric Myhre <hash@exultant.us>