Begin amending json test logic to support error-path assertions.
But this is a real mess.  These outputs are utterly awful to read and I'm... not gonna anymore.  For fuck's sake, the output doesn't even clearly say which lines are logs and which are errors; just FAIL, $n lines above, with all the clarity of mud.  We already converted all the object marshal/unmarshal to GoConvey; let's do the same here momentarily.

Signed-off-by: Eric Myhre <hash@exultant.us>