Fix null pointer stomping existing data due to bug in ptr defer unmarshal not delegating hard enough after first steps.
Regression test coverage included... though it took me quite a while to figure out what exactly was required to cover it: the pointers have to be in a multi-field struct, and they can't be in the top level one (unless you hand in a pointer to its pointer), AND the null has to come late in the struct, or you don't hit the critical path.  Woo.

Signed-off-by: Eric Myhre <hash@exultant.us>