obj: initialize slices unconditionally on unmarshal.
Fixes a bug where a slice unmarshaller being reused by a map unmarshaller would leak increasing sizes of slice due to nonsensical reuse of a temp slice handle.

This branch was an attempt to save allocations if given a struct that already had a bunch of arrays, but it's doesn't live up to that task very effectively.  Attempting to get that optimization right requires better test coverage and much better benchmarks targetted specifically at it before undertaking.

Signed-off-by: Eric Myhre <hash@exultant.us>