json: parsing ints successfully.
Our big ol' rebase actually saved a bunch of strife: the gentler EOF handling from the extracted SlickReader util is essential, and the error returns from the read functions instantly pointed out that this for-loop's step component was trying to do *way* too much at once.

We still need more state machinery to parse actual floats.  That should just be a plug and chug sort of operation.

I hacked the type of token yielded to force it to int.  This is obviously wrong and needs a fix.  Which is... an interesting subject.

We're probably going to attempt to cast things to ints if we can do so.  Our object unmarshaller is perfectly happy to take ints and coerce them into a float-type field, but it's (rightly) not willing to silently do the converse and coerce a float down into an int field.  Json's typeless inanity here shall remain the json library's problem: all truly bonkers casts and range checks will stay in the json package, and not be allowed to ruffle the obj package.

Signed-off-by: Eric Myhre <hash@exultant.us>