cbor,tok: retain 'tag' values in Token field.
We use two fields in the Token struct.  This is to resolve ambiguity with cbor tags of the int value zero.  An earlier iteration of the design tried to get away with one field, and use -1 as a sentinel value, but this generated a *lot* of implementation pain, because that sentinel value would need initialization e v e r y w h e r e in the entire project's codebase, whereas the use of tags is confined to a very small area.  It's not worth optimizing for the extra bytes in struct layout; we reuse the bejezus out of the token struct already anyway.

The json decoder chooses *not* to bother to re-false the Token.Tagged field -- it just believes that you'll hand it a roughly zero'd token, and not inject screwyness.  I think this is sensible and fine given that all the actual user-facing methods for initializing full marshal/unmarshal systems do in fact allocate one plain Token, and it's never reused with completely different implementations of decoder logic which would actually give that field a way to become set.

Partially fixes #7.

Signed-off-by: Eric Myhre <hash@exultant.us>