link a yaml library
... god have mercy on my soul.

Yaml is a trashfire of underspecified (or, overspecified-but-nonsense) mess.  I will not attempt to write a parser for it because some of the best minds of our generation have looked at the original specs and parsers and wept trying to port them.  It's not clear if ruby's parse.y or the yaml grammar is worse.  They are so bad they have been subjected to *reverse engineering* and the brave souls going on such quests have *reported the expedition a failure* despite nothing short of heroism.

So.  I am not writing a new yaml parser.

But I will happily use this library.

Signed-off-by: Eric Myhre <hash@exultant.us>