obj: unmarshal can now be directed by tags!
Currently, this only works when the unmarshaller has no type info at all; then we'll fabricate a value based on the tag.  In all other paths, we're quite permissive (read: don't notice at all) if tags are present in the token stream but would indicate other types than the ones we have at hand in the marshaller.  We may get more strict about this in the future; this feature is still in its early stages.

Tags are currently not emitted by the marshaller at all.

Test coverage exists for the basics of tag unmarshalling -- you can see it being used to emit a type when a wildcard would have otherwise resulted!

Signed-off-by: Eric Myhre <hash@exultant.us>