obj: tags are now supported on transforms, both ways.
I'm not very pleased with the shape of some of this code.  The presense of *data controlled* behavior rather than type-info controlled is simply two different philosophies butting up against each other, and there's an upper limit to how graceful it can be, so some ugliness is to be expected.  (But this could still be better :/  Refactors in the future, perhaps.  I'm starting to wrack up quite a list of frustrations with the current entwinement of slabs and atlases, which is potentially related.)

Signed-off-by: Eric Myhre <hash@exultant.us>