cbor: fixes to nil encoding.
Two sources of problem:

- the wrong byte was being encoded for nil...!  Youch.
- a stack pop was performed as if nils weren't terminals, which could cause panics.

Both should now be fixed.

The fixtures we already had for empty and zero values should've caught this... but unfortunately we hadn't mapped them onto corresponding CBOR byte fixtures.  That oversight is now fixed (and we added more fixtures, to boot), but we should also really fix the root cause of that oversight and write a test that we've got enough tests mapped consistently in all directions >.<

Thanks to dignifiedquire for his stellar bug reporting.  This should fix issue #10.  (The "invalid major byte" from unmarshal was actually a valid complaint; it's also fixed by fixing the wrong byte emitted from the marshal side.)

Signed-off-by: Eric Myhre <hash@exultant.us>