refmt: add more tests, focusing on roundtripping.
These are fairly overlapping with the combined coverage of unit tests in the individual serial formats and the object package... but putting them together in one place for peace of mind, clarity, and simplicitly.

So far I haven't made anything fail.  (These tests are particularly trying to make sure we cover any nil-vs-typed-nil issues, because that's often a particularly pernicious thing in Go.)

Signed-off-by: Eric Myhre <hash@exultant.us>