Refrain from exporting the altas.Atlas.MapMorphism field.
Prefer to use accessor methods so we're at least that step closer to making
it immutable.

Of course it's not properly immutable, and we sort of give up because golang
is not a terribly powerful language and we've already long admitted in this
project that we'll do unsafe things for performance.
(That said, I also would really like to have some time to invest into
benchmarking some of these details again, possibly with the aim of removing
more pointer faffery.)

Remove a case in a switch that I'm fairly sure didn't belong there -- that
switch as a whole is only hit if you're *got* a specific entry of config,
so a branch in it that considers using the default config is nonsensical.
I'd make that change in a separate commit, but keeping it across this
change to an accessor method would make a mess itself.

Also, atlas building now always sets the default mapMorphism.  So, we
can chase down a bunch of unnecessary/redundant nil pointer checks shortly.

Signed-off-by: Eric Myhre <hash@exultant.us>