Fix several missing slab.release calls in obj.Unmarshal*.
The lack of these wouldn't cause incorrectness, but it would cause the amount
of memory used by a single unmarshaller run to be excessive; the slab would
grow more than necessary because rows wouldn't be reclaimed all the time.

Manual memory management is a joy.

All the 'requisition' call sites are now properly paired with a 'release' call.
There's an exception around error cases, but those don't matter because the
only next thing that can be done with an errored unmarshaller machine stack
is reset it completely, which zeros the slab length anyway.

Signed-off-by: Eric Myhre <hash@exultant.us>