Fix and regression test for map-struct-map bug.
This is a nasty correctness bug that only strikes when you have maps nested
indirectly inside other maps, due to a convergence between reuse of the
"temp" slot we need for map unmarshalling (which is in turn because reflect
can't operate on unaddressable handles, which map values *are*) and the
branch on initializing the target in map unmarshal which is intented to make
it possible to reuse existing allocated maps if given them.  Woof.

This increases the amount of heap memory allocations we have to make during
map unmarshalling, which is unfortunate.  I'm going to try to improve on that
without breaking the regression test in a moment.

Signed-off-by: Eric Myhre <hash@exultant.us>