First pass of omitEmpty support.
Structs now inspect the value before each key, because yielding of the key
must of course be skipped if the value is to be skipped.

And yet, we're not done here, and that test is commented out for a reason!
This is more complicated than for e.g. stdlib json.Marshal -- we have to
emit length information at the beginning of an object.

And this, in turn, is capital-H Hard.  Emitting the correct length information
*up front* will require significantly more code changes, and they're a tad
controvertial.  We have to inspect *all* the fields to see if they're going
to be skipped.  And strangely, I think we're going to have to do that twice.
Checking for the fields to skip must happen at the top, that much is clear;
but then to remember which ones we already know will be skipped would require
O(n) memory in the length of the struct... which would imply a heap allocation
to track!  (Worrying about heap allocs is not news in the refmt project because
of our stepfunc design, but it's interesting to note we'd be in trouble anyway:
Go actually always lets runtime-sized slice creation escape to heap:
https://github.com/golang/go/issues/20533 .)  So.  An O(2n) runtime is going
to be a better trade than slipping from constant to O(n) memory.  Hrmph.
Anyway, that bit will be in the next commits.

Signed-off-by: Eric Myhre <hash@exultant.us>