OmitEmpty now operating correctly.
The test fixture for emitting an empty (and zero length!) map from a struct
with only omitEmpty and emptied fields now passes.

As discussed in the previous commit, fully correct omitEmpty support requires
counting how many fields will be emitted up front: this patch does so.

The checks for emptiness are done twice: once up front to produce the correct
length header, and again as we iterate over the object.
This means our time complexity went from O(n) to O(2n) -- however, it keeps
memory needed to traverse the struct at O(1) no matter how many fields the
struct has; other algorithms that try to reduce time costs by remembering
the emptiness of the fields ends up requiring *memory proportial to the
number of fields in the struct*, and that's a significantly worse trade.

Signed-off-by: Eric Myhre <hash@exultant.us>