Restructure slab usage by struct marshalling.
Since we don't need to do a machine for fields we're going to skip, this needed
review... but it turns out we actually don't need to be faffing about with
shrinking and growing the slab for each field at all; we can just keep reusing
that row... so, correctness, and now a wee performance boost to boot.  Yay.

Signed-off-by: Eric Myhre <hash@exultant.us>