testing: very first case ported to go-wish.
holy wow this is better.  In the case of errors, you'll now get a reasonably
formatted and golang-community-standard logf with source file and line number
that actually points directly to the fixture that's problematic.

Also, all fuckery with title composition is now handled by the t.Run name
parameters, and that's pretty fine.

The output is much better and much easier to attain because now all of the
assertion logic is based on "gather all data; THEN compare".  No branching.

It doesn't show up yet in this small example, but when we get to the obj
package in particular, having local scopes (something not possible with
tests coded up as purely table-driven) for declaring more struct types
is going to be a drastic sanity enhancer for reducing how much the reader
has to scroll around.  (We're still effectively *doing* table-driven
test style, semiotically -- we're going to end up with massive "tables"
of `t.Run ... checkEncoding` -- but doing it with funcs rather than
declarative struct data means we can have scoping.  And that's Nice in
the domain of the obj marshal/unmarshal tests.)

I'm going to miss the nice green checkmarks from go-convey.  But the
overall benefits of this conversion are just massive.

Signed-off-by: Eric Myhre <hash@exultant.us>