testing: second case ported to go-wish.
See parent commit for more commentary on how this port is awesome.

Again, if you compare this new check function to the one previously in
jsonDecoder_test.go (not seen in this diff, unfortunately, because
we're not removing it until the port is complete), it's pretty clear
that being able to accumulate your total results and do a single
assertion at the end results in radically simpler test code.
Look Ma, no branching!

Doing the Wish(...) on the token array as whole also provides pretty
great output.  Single-token errors will be reported as such, tersely.
I don't think there's any detection for shifts, unfortunately.
But we'll see if that's a problem in practice (I suspect not; the human
can eyeball that kind of feedback pretty easily), and if it is...
we have the tools to counter it: we'll write a wish.Checker func that
decorates ShouldEqual to stringify the token slices first, and then
the string diffing algos will kick in and give patch-style diffs.

Signed-off-by: Eric Myhre <hash@exultant.us>