testing: add 'checkCanonical' helper.
This gets rid of a lot of copypasta in the extremely common case that
there's a canonical serial form and so of course both encode and
decode should both handle it.

There are some token sequences which *have* no canonical form and
thus won't have this check, but that's okay too.

Signed-off-by: Eric Myhre <hash@exultant.us>