testing: migrate cbor bool testcases to go-wish.
Refraining from the 'checkCanonical' pattern used in the json package.
Turns out that approach has some issues.  The go testing tools only
regard your declarations of `t.Helper` up until they encounted your
`t.Run` func in the stack... which means if you have a helper method
calling `t.Run` itself, GAME OVER: you CANNOT have it report line
numbers in `Logf` calls that point to the callsite above that helper.

This pretty much leaves us with the same major complaint we previously
had when doing table-driven tests: the line numbers reported for
failures were basically useless, because the callsite is always the
same, and all the interesting and semantically relevant data was in a
table somewhere totally unrelated.

(I both see the reasoning there, and yet am not pleased with it,
because generating fairly standardly-shaped trees of test funcs
and applying them is something I do *often* for interface conformance
tests, for example, and I kinda want that experience to not suck.)

Signed-off-by: Eric Myhre <hash@exultant.us>