obj: introduce new testing with go-wish.
I'm going to refrain from trying to pull off a single giant refactor
in this package.  In part, this is because it's simply *daunting*:
we have around 1.5k assertions being made in the existing framework.
But also, because taking things slowly will provide a pleasant
opportunity to review our coverage of all the paths; the new code has
noticably different ergonomics that'll make many cases easier.

Note how in the added tests here, we got a ton more coverage of what
happens when you try to map the same sequence onto different structs;
and it's very readable.  Because we can write types in-line, with
both limited scope and sheer proximity to application, object mapping
tests against a variety of types are vastly easier to write now.

Signed-off-by: Eric Myhre <hash@exultant.us>