obj: First pass on union support for interfaces!
This is an interesting and tricky thing.  As far as I'm aware, no other
Golang serialization libraries have features comparable to this, so in
addition to just plain implementation fun, we're going to be blazing
trails on UX of the API as well.  Here we go!

(Several TODOs are sneaking in.  This is tricky enough that iterative
progress over many commits seems necessary.)

Signed-off-by: Eric Myhre <hash@exultant.us>