obj: union support for interfaces now round-trips!
Diff contains a couple of odd random fixups to the marshal system
overall.  Unmarshalling had some conveniences like 'slab.tip' and some
more function interface types that I have no idea how or why the
marshal side was missing.

The 'rtid' computation is starting to show up in more and more places,
and that's getting a tad alarming.  A refactor to clean that up and
give it a more reified single home in the code would be good.
I wonder if perhaps it should be cached and passed down in more places
in parallel to the various 'foo_rt' values; I haven't reviewed in
detail, but I suspect we're computing it redundantly sometimes.

Signed-off-by: Eric Myhre <hash@exultant.us>