obj: detect typedef'd fixed-length byte arrays.
Previously, fixed-length typedef'd byte arrays would be encoded
arguably *correctly*, but not densely: the array would serialize
in e.g. cbor as roughly twice as many bytes as it should, because
the byte type indicator would be prepended to each byte individually,
rather than the whole array marked as bytes at once.  Now, these
types will be turned into tokens of TBytes correctly, and thus
serialize appropriately densely.

This is... a surprisingly deeply different path.  The reflect package
does not treat slices and fixed-arrays *at all* similarly.

Added new tests in the new form.  (A bit sad that past self hasn't
finished churning through conversion of all the existing obj tests
yet.)

This is just the marshal side; unmarshal coming up.

I'm not pleased with the efficiency nor elegance of this; see comments
in the diff.  It seems fixed-length arrays are simply a second-class
citizen in Go, especially in the reflect package.  ヽ(´ー｀)ノ
Or, I'm missing something obvious (which I would love to be the case,
and someone out there, please kick me in the shins when you find a
better solution for this).

Signed-off-by: Eric Myhre <hash@exultant.us>