obj: unmarshalling into fixed-length byte arrays.
Same thing as with marshalling: there's a loop in here, and it's not
pretty, but I'm at a loss for any cleaner more efficient way to do this
given the constraints of the reflect package API.  PRs welcome.

This is a large diff because I added a field to an existing error type,
and this caused changes at the many, many initializer sites.

Signed-off-by: Eric Myhre <hash@exultant.us>