bench: fix obj test; make corresponding map test.
Had to rearrange the obj fixtures and types somewhat, because with maps
and stdlib json, there's no way to define non-canonical orders -- which
our obj fixtures were previously exercising.

It's a nice little example of how much faster struct operations can be
than maps -- a *lot*, turns out, given that you have to deal with key
sorting, and that forces both the refmt and the stdlib implementations
into quite a few memory allocations.

Signed-off-by: Eric Myhre <hash@exultant.us>