cbor: introduce CoerceUndefToNull decode option.
By default we will raise an error if we encounter the cbor "undefined"
byte, because it's a long way from a canonical object if it has one of
these bytes in it, and I think we should not be silent about that.
If you want to coerce it to null: that's also an option!

Diff is a bit larger than one might expect it should be because this
turns out to be the first decoding option we've actually passed
through all the way to the decoder.

Signed-off-by: Eric Myhre <hash@exultant.us>