Interesting new exploratory microbenchmarks.
Some of these have been laying around a while, uncommitted (tsk).
The BenchmarkBoxingPrimitives set are from today, and are particularly
interesting.

And wow.  Today I learned some things.

Also, I learned that a bunch of this stuff changed massively for the
better in the compiler and runtime core.... riiiight around the time
refmt was getting in motion..... being developed on the old compilers
and runtimes in their last days before the new stuff tagged.  Ow.

Signed-off-by: Eric Myhre <hash@exultant.us>