Yank library updates.
In particular, a bump to gls was needed, which addresses
https://github.com/smartystreets/goconvey/issues/476 .

(I'd like to finish moving from GoConvey to go-wish, which uses a
more explicit approach to test trees which became more natural
after the stdlib introduced t.Run, and thus dodges any potential
of arcane issues like this.  But that's a "not today" priority.)

Skipped yanking updates of the yaml libs because it gave some fuff
about import paths and again that's just "not today", thanks.

Signed-off-by: Eric Myhre <hash@exultant.us>