Add several new benchmarks.
There is considerable yak-shaving to be seen in the area of the
"fixFloats" function.  I... don't even want to talk about it.

It turns out I previously skipped over benchmarking these particular
paths because (I remember now, in horror) of these issues of
inconsistency in how different libraries handle numerics... which made
it painfully irritating to use the same fixtures... which...

Boy, wouldn't you know.  This one part that we didn't benchmark:
it's *wildly* slower and more memory alloc hungry than anything else
in the whole library.

Ouch.

Ouuuuuuchhhhh.  Ouchouch.

Fixes will be coming soon.  (I've already staged them; just want to
land the benchmarks first so that it's easier to get a readout of just
how big the improvements are.)

Switched several asserts to use go-wish/go-cmp along the way.
Nicer output in case of errors.  (Extra costs still get swallowed as
long as your benchmark run count is high enough -- and the default one
second goal still results in runcounts well above the hundreds of
thousands, so, yeah, the noise is indeed drowned in signal.)