Unmarshalling without type info prefers int64.
Similar changes were commit 36d31e7716ec3a8517035447a57dde1d88cd528e;
but there, we used 'int', with the reasoning "Plainer is better."
I've since received complaints about this: that on 32-bit architectures
this will result in different behavior than on 64-bit, and that it may
be lossy on 32-bit.

Since the only downside of using int64, as far as I know, is that it
takes a few more characters on the source side -- 'int' is already uses
64-bits of space on a 64-bit architecture anyway -- I guess I'll
concede that it might as well be the default.

As before, it's important to address this at all because otherwise we
would experience some odd behaviors in that the CBOR unmarshal and
JSON unmarshal would behave differently when targeting an empty
interface: JSON would yield int64, while CBOR would yield *uint64*,
due to CBOR's Interesting (ahem) Opinions about signedness of numbers.

And as before, remember: if unmarshalling into a concrete type that
uses some other numeric type like `uint64`, etc, those unmarshallings
do not pass through this cast and are not affected by this change.